migrated code to nested routing
